Moved custom sensor exceptions to utils.exceptions
